api/pyproject.toml
[project]
name = "auth-api"
version = "0.1.0"
description = "Authentication API with FastAPI"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "bcrypt>=4.1.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "slowapi>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true